Align API to changes from IISNative
JIRA: AO-17842

+++ v2/bindings/csharp/SnapPluginLibCs/SnapPluginLib/CollectContext.cs
@@ -11,7 +11,7 @@ namespace SnapPluginLib
         {
         }
 
-        public void AddMetric(string ns, int value, params Modifier[] modifiers)
+        public void AddMetric(string ns, long value, params Modifier[] modifiers)
         {
             var nativeValue = new NativeValue
             {
