+++ v2/bindings/csharp/SnapPluginLibCs/SnapPluginLib/ICollectContext.cs
@@ -5,7 +5,7 @@ namespace SnapPluginLib
     public interface ICollectContext : IContext
     {
         void AddMetric(string ns, double value, params Modifier[] modifiers);
-        void AddMetric(string ns, int value, params Modifier[] modifiers);
+        void AddMetric(string ns, long value, params Modifier[] modifiers);
         void AddMetric(string ns, uint value, params Modifier[] modifiers);
         void AddMetric(string ns, bool value, params Modifier[] modifiers);
         void AddMetric(string ns, string value, params Modifier[] modifiers);
